relocate endpoints to smithy-client
