+++ runtime/serde/common/src/aws/smithy/kotlin/runtime/serde/Serializer.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 package aws.smithy.kotlin.runtime.serde
-import aws.smithy.kotlin.runtime.smithy.Document
+import aws.smithy.kotlin.runtime.content.Document
 import aws.smithy.kotlin.runtime.time.Instant
 import aws.smithy.kotlin.runtime.time.TimestampFormat
