+++ runtime/smithy-client/common/src/aws/smithy/kotlin/runtime/client/endpoints/Endpoint.kt
@@ -3,16 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package aws.smithy.kotlin.runtime.http.endpoints
+package aws.smithy.kotlin.runtime.client.endpoints
 
 import aws.smithy.kotlin.runtime.InternalApi
-import aws.smithy.kotlin.runtime.http.Headers
-import aws.smithy.kotlin.runtime.http.operation.HttpOperationContext
-import aws.smithy.kotlin.runtime.http.operation.SdkHttpRequest
-import aws.smithy.kotlin.runtime.net.Host
 import aws.smithy.kotlin.runtime.net.Url
 import aws.smithy.kotlin.runtime.util.AttributeKey
 import aws.smithy.kotlin.runtime.util.Attributes
+import aws.smithy.kotlin.runtime.util.ValuesMap
 
 /**
  * Represents the endpoint a service client should make API operation calls to.
@@ -39,7 +36,7 @@ import aws.smithy.kotlin.runtime.util.Attributes
  */
 public data class Endpoint @InternalApi constructor(
     public val uri: Url,
-    public val headers: Headers = Headers.Empty,
+    public val headers: ValuesMap<String>? = null,
     @InternalApi
     public val attributes: Attributes = Attributes(),
 ) {
@@ -47,7 +44,7 @@ public data class Endpoint @InternalApi constructor(
 
     public constructor(
         uri: Url,
-        headers: Headers = Headers.Empty,
+        headers: ValuesMap<String>? = null,
     ) : this(uri, headers, Attributes())
 
     override fun equals(other: Any?): Boolean =
@@ -62,34 +59,4 @@ public data class Endpoint @InternalApi constructor(
                 @Suppress("UNCHECKED_CAST")
                 attributes.contains(it) && attributes.getOrNull(it as AttributeKey<Any>) == other.attributes.getOrNull(it)
             }
-}
-
-/**
- * Update an existing request with a resolved endpoint.
- *
- * Any values serialized to the HTTP path or query string are preserved (in the case of path, the existing serialized one
- * is appended to what was resolved).
- */
-@InternalApi
-public fun setResolvedEndpoint(req: SdkHttpRequest, endpoint: Endpoint) {
-    val hostPrefix = req.context.getOrNull(HttpOperationContext.HostPrefix) ?: ""
-    val hostname = "$hostPrefix${endpoint.uri.host}"
-    val joinedPath = buildString {
-        append(endpoint.uri.path.removeSuffix("/"))
-        if (req.subject.url.path.isNotBlank()) {
-            append("/")
-            append(req.subject.url.path.removePrefix("/"))
-        }
-    }
-
-    req.subject.url.scheme = endpoint.uri.scheme
-    req.subject.url.userInfo = endpoint.uri.userInfo
-    req.subject.url.host = Host.parse(hostname)
-    req.subject.url.port = endpoint.uri.port
-    req.subject.url.path = joinedPath
-    req.subject.url.parameters.appendAll(endpoint.uri.parameters)
-    req.subject.url.fragment = endpoint.uri.fragment
-
-    req.subject.headers["Host"] = hostname
-    req.subject.headers.appendAll(endpoint.headers)
 }
