+++ runtime/smithy-client/common/src/aws/smithy/kotlin/runtime/client/endpoints/EndpointProviderException.kt
@@ -2,7 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package aws.smithy.kotlin.runtime.http.endpoints
+package aws.smithy.kotlin.runtime.client.endpoints
 
 import aws.smithy.kotlin.runtime.ClientException
