+++ runtime/smithy-client/common/test/aws/smithy/kotlin/runtime/client/endpoints/functions/FunctionsTest.kt
@@ -2,7 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-package aws.smithy.kotlin.runtime.http.endpoints.functions
+package aws.smithy.kotlin.runtime.client.endpoints.functions
 
 import kotlin.test.*
